requirements.txt
# Core dependencies
numpy==1.26.4
pandas==2.0.3
scikit-learn==1.3.0
scipy==1.10.1
torch==2.0.1
transformers==4.36.0
sentence-transformers==2.2.2
flask==3.0.0
pyyaml==6.0.1
python-dotenv
huggingface-hub==0.19.3

# NLP for compound noun extraction (optional - has heuristic fallback)
spacy>=3.0.0,<4.0.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl

# LLM APIs
openai>=1.0.0
google-generativeai>=0.3.0

# Web interface
jinja2>=3.0.0
flask-cors>=4.0.0

# Formal methods
z3-solver>=4.8.0
pyModelChecking>=1.3.3
rtamt

# Development
pytest>=7.0.0
tqdm>=4.65.0

# Optional
ai2thor>=5.0.0
